third_party/WebKit/LayoutTests/fast/html/imports/sub-imports.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<link id="root" rel="import" href="resources/root.html">
</head>
<body>
<script>
shouldBe("root.import.querySelector('h1').innerHTML", "'Root'");

hello = root.import.querySelector('#hello');
shouldBe("hello.import.querySelector('h1').innerHTML", "'Hello'");
bye = root.import.querySelector('#bye');
shouldBe("bye.import.querySelector('h1').innerHTML", "'Bye'");

</script>
</body>
</html>
