third_party/WebKit/LayoutTests/fast/html/imports/sub-imports-expected.txt

PASS root.import.querySelector('h1').innerHTML is 'Root'
PASS hello.import.querySelector('h1').innerHTML is 'Hello'
PASS bye.import.querySelector('h1').innerHTML is 'Bye'
PASS successfullyParsed is true

TEST COMPLETE
